fix(onboarding): render signup hero from the real V1 funnel step
The previous hero implementation lived behind an unreachable gate in
pages/onboarding.tsx (`isAuthenticating && defaultDisplay ===
OnboardingSignup`). In the live V1 product the signup screen is rendered
by FunnelOrganicSignup inside FunnelStepper, so the hero was effectively
dead code — visitors of /onboarding never reached it.

- Move OnboardingSignupHero into shared (features/onboarding/components)
  so the funnel step can consume it, replace styled-jsx with a plain
  <style> tag, and add an `isFormExpanded` prop that collapses the
  marketing surface when the email/social registration form takes over.
- Swap FunnelOrganicSignup's render with <OnboardingSignupHero>,
  preserving every auth callback (onAuthStateUpdate,
  onSuccessfulRegistration, onSuccessfulLogin) and the existing
  isLoggedIn / infoConfirmed / isOnboardingComplete guards. Force dark
  theme on mount; restore on unmount.
- Drop the unreachable signup branch and unused imports
  (ThemeMode, AuthDisplay.OnboardingSignup theming effect,
  OnboardingSignupHero dynamic import) from pages/onboarding.tsx.
- Add FunnelOrganicSignup.tsx to the strict-typecheck skip list,
  matching the precedent for the other auth/funnel files (the surfaced
  strict errors are pre-existing on unchanged logic copied from the
  original step).
