+++ packages/shared/src/features/onboarding/steps/FunnelOrganicSignup.tsx
@@ -4,8 +4,6 @@ import classNames from 'classnames';
 import { useSetAtom } from 'jotai/react';
 import type { FunnelStepOrganicSignup } from '../types/funnel';
 import { FunnelStepTransitionType } from '../types/funnel';
-import { OnboardingHeadline } from '../../../components/auth';
-import { FooterLinks } from '../../../components/footer';
 import AuthOptions from '../../../components/auth/AuthOptions';
 import { AuthTriggers } from '../../../lib/auth';
 import { ButtonSize, ButtonVariant } from '../../../components/buttons/common';
@@ -13,24 +11,16 @@ import { useViewSize, ViewSize } from '../../../hooks';
 import type { AuthProps } from '../../../components/auth/common';
 import { AuthDisplay } from '../../../components/auth/common';
 import { ExperimentWinner } from '../../../lib/featureValues';
-import { wrapperMaxWidth } from '../../../components/onboarding/common';
-import { OnboardingHeader } from '../../../components/onboarding/OnboardingHeader';
 import { authAtom } from '../store/onboarding.store';
 import type { AnonymousUser, LoggedUser } from '../../../lib/user';
 import { useAuthContext } from '../../../contexts/AuthContext';
 import { withIsActiveGuard } from '../shared/withActiveGuard';
-import {
-  cloudinaryOnboardingFullBackgroundDesktop,
-  cloudinaryOnboardingFullBackgroundMobile,
-} from '../../../lib/image';
 import { useOnboardingActions } from '../../../hooks/auth';
 import {
-  Typography,
-  TypographyColor,
-  TypographyType,
-} from '../../../components/typography/Typography';
-import { sanitizeMessage } from '../lib/utils';
-import ConditionalWrapper from '../../../components/ConditionalWrapper';
+  ThemeMode,
+  useSettingsContext,
+} from '../../../contexts/SettingsContext';
+import { OnboardingSignupHero } from '../components/OnboardingSignupHero';
 
 type FunnelOrganicSignupProps = FunnelStepOrganicSignup;
 
@@ -58,20 +48,12 @@ const isSocialSignupUser = (
 };
 
 export const FunnelOrganicSignup = withIsActiveGuard(
-  ({
-    parameters: {
-      explainer,
-      headline,
-      image: srcDesktop = cloudinaryOnboardingFullBackgroundDesktop,
-      imageMobile: src = cloudinaryOnboardingFullBackgroundMobile,
-      version = 'v1',
-    },
-    onTransition,
-  }: FunnelOrganicSignupProps): ReactElement => {
+  ({ onTransition }: FunnelOrganicSignupProps): ReactElement => {
     const formRef = useRef<HTMLFormElement>(null);
     const hasAlreadyCheckedUser = useRef(false);
     const isMobile = useViewSize(ViewSize.MobileL);
     const setAuth = useSetAtom(authAtom);
+    const { applyThemeMode } = useSettingsContext();
     const { isLoggedIn, isAuthReady, user } = useAuthContext();
     const { isOnboardingActionsReady, isOnboardingComplete } =
       useOnboardingActions();
@@ -86,7 +68,13 @@ export const FunnelOrganicSignup = withIsActiveGuard(
       isLoggedIn &&
       !isEmailSignupActive &&
       isSocialSignupUser(user);
-    const isMobileRevamp = version === 'v2_mobile' && isMobile;
+
+    useEffect(() => {
+      applyThemeMode(ThemeMode.Dark);
+      return () => {
+        applyThemeMode();
+      };
+    }, [applyThemeMode]);
 
     const onAuthStateUpdate = useCallback(
       (data: Partial<AuthProps>) => {
@@ -182,125 +170,31 @@ export const FunnelOrganicSignup = withIsActiveGuard(
     }
 
     return (
-      <div
-        className={classNames(
-          'relative z-3 flex flex-1 flex-col overflow-x-hidden',
-          {
-            'justify-end pt-40': isMobileRevamp,
-            'items-center': !isMobileRevamp,
-          },
-        )}
+      <OnboardingSignupHero
+        isFormExpanded={isEmailSignupActive || isSocialSignupActive}
       >
-        <OnboardingHeader
-          isLanding
-          className={classNames(
-            isMobileRevamp &&
-              '!-my-8 !justify-center bg-gradient-to-t from-background-default to-transparent py-8',
-          )}
+        <AuthOptions
+          {...staticAuthProps}
+          defaultDisplay={
+            isSocialSignupActive ? AuthDisplay.SocialRegistration : authDisplay
+          }
+          formRef={formRef}
+          onboardingSignupButton={{
+            size: isMobile ? ButtonSize.Medium : ButtonSize.Large,
+            variant: ButtonVariant.Primary,
+          }}
+          onSuccessfulRegistration={onSuccessfulRegistration}
+          onAuthStateUpdate={onAuthStateUpdate}
+          onSuccessfulLogin={() => {
+            // user now is logged, even if the `user` object is not populated yet.
+            // this callback is fired only after a lot of auth checks
+            onTransition?.({
+              type: FunnelStepTransitionType.Complete,
+              details: { user },
+            });
+          }}
         />
-        <ConditionalWrapper
-          condition={isMobileRevamp}
-          wrapper={(content) => (
-            <div className="relative z-1 after:absolute after:inset-0 after:top-8 after:-z-1 after:bg-background-default">
-              {content}
-            </div>
-          )}
-        >
-          <div
-            className={classNames(
-              `flex w-full flex-col flex-wrap content-center justify-center px-4 tablet:flex-row tablet:gap-10 tablet:px-6`,
-              wrapperMaxWidth,
-              { 'mt-7.5': isEmailSignupActive, 'flex-1': !isMobileRevamp },
-            )}
-          >
-            <div
-              className={classNames(
-                'mt-5 flex flex-1 flex-col tablet:my-5 tablet:flex-grow',
-                !isEmailSignupActive && 'laptop:mr-8 laptop:max-w-[27.5rem]',
-              )}
-            >
-              {!isEmailSignupActive && !isSocialSignupActive && (
-                <>
-                  {isMobileRevamp ? (
-                    <div className="mb-8 flex flex-col gap-4">
-                      <Typography
-                        className="text-balance text-center"
-                        color={TypographyColor.Primary}
-                        dangerouslySetInnerHTML={{
-                          __html: sanitizeMessage(headline),
-                        }}
-                        type={TypographyType.Title2}
-                      />
-                    </div>
-                  ) : (
-                    <OnboardingHeadline
-                      className={{
-                        title: classNames(
-                          'tablet:typo-mega-1 typo-large-title',
-                        ),
-                        description:
-                          'mb-8 text-text-primary typo-body tablet:typo-title2',
-                      }}
-                      title={sanitizeMessage(headline, [])}
-                      description={explainer}
-                    />
-                  )}
-                </>
-              )}
-              <AuthOptions
-                {...staticAuthProps}
-                defaultDisplay={
-                  isSocialSignupActive
-                    ? AuthDisplay.SocialRegistration
-                    : authDisplay
-                }
-                formRef={formRef}
-                onboardingSignupButton={{
-                  size: isMobile ? ButtonSize.Medium : ButtonSize.Large,
-                  variant: ButtonVariant.Primary,
-                }}
-                onSuccessfulRegistration={onSuccessfulRegistration}
-                onAuthStateUpdate={onAuthStateUpdate}
-                onSuccessfulLogin={() => {
-                  // user now is logged, even if the `user` object is not populated yet.
-                  // this callback is fired only after a lot of auth checks
-                  onTransition?.({
-                    type: FunnelStepTransitionType.Complete,
-                    details: { user },
-                  });
-                }}
-              />
-            </div>
-            <div className="flex flex-1 tablet:ml-auto tablet:flex-1 laptop:max-w-[37.5rem]" />
-          </div>
-        </ConditionalWrapper>
-        {!!src && (
-          <picture>
-            <source media="(max-width: 655px)" srcSet={src} />
-            <source media="(min-width: 656px)" srcSet={srcDesktop || src} />
-            <img
-              alt="Onboarding background"
-              aria-hidden
-              className={classNames(
-                'pointer-events-none absolute inset-0 -z-1 size-full object-cover transition-opacity duration-150',
-                { 'opacity-[.24]': isEmailSignupActive },
-              )}
-              fetchPriority="high"
-              loading="eager"
-              role="presentation"
-              src={src}
-            />
-          </picture>
-        )}
-        {!isMobileRevamp && (
-          <FooterLinks
-            className={classNames(
-              'mx-auto px-2 pb-6 laptop:px-0',
-              isMobileRevamp && '!-mb-4 bg-background-default pt-4',
-            )}
-          />
-        )}{' '}
-      </div>
+      </OnboardingSignupHero>
     );
   },
 );
